Revert "Add obj/ folders to projects"
This reverts commit 5ae65d2b4ae08ab4fa6883e24fcd9a0fc286a80f.
Apparently we *actually don't* need any of this obj shit.
It is indeed temporary build garbage that the .NET compiler
will generate when you build it.

I thought we needed these files because a teammate was having
problems building/running from Visual Studio, but using the
dotnet CLI worked just fine, so it was a problem related to a
misconfigured VS environment of some kind.