Integrate Website and Api processes to run together
Api/Program.cs needed some services added from Microsoft.AspNetCore.Components.WebAssembly.Server
in order to launch the Blazor frontend project along with it.

With these changes, you can now launch the Api process, and it will
automatically build and launch the Blazor project, as well as watch
the Blazor files for updates and hot reloads.  In VSCode (or in the
command line), you can run
`dotnet watch run`
in the Api project directory, to automatically launch and watch all
frontend and backend project files for updates, as well as enable
hot reloading.  Additionally, if a "rude edit" prevents dotnet
from hot reloading, it will prompt you for a rebuild: enter "a"
to always rebuild, and the development process becomes that much
easier.