Add database connections
Including a connection string in plain text in version
control is a stupid thing to do but I did it anyway
because I wanted to prove that there's really a postgres
database out there somewhere in the great beyond.

This database user is short-lived and will be annihilated
when we have an actual deployment pipeline set up, and
instead there will be an environment variable on the
deploy host so that nobody will have to know about the
connection string at all.