Make test of `AuthController.Register()` actually work
Firstly, the `appsettings.json` file is available in the Tests project's
executing directory, since the project reference to `Api` packs all the
necessary dependencies into bin/Debug.

Secondly, testing that the return value of `AuthController.Register()` is
not null is meaningless, because a `Task<T>` is always non-null.  The task
must be waited, and the contents of the result inspected for any test of
this method to be meaningful.

Thirdly, the data context was not setup properly: none of the tables
existed.  This was addressed by calling
`_dbContext.Database.EnsureCreated()`.

Fourth, the data context shouldn't be freed at the end of one test, but
upon the disposal of the testing class: therefore the call to
`_connection.Dispose()` was moved to the definition of
`AuthControllerTests.Dispose()`.

Fifth, the test was invalid because one of the test payload's required
properties was left unset: `Avatar` was required.

Finally, all of the nested type parameters returned from
`AuthController.Register()` were properly checked and unwrapped, all the
way down to the *actual* token.